requirements.txt
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
idna==3.11
ipython==8.38.0
ipython-beautifulsoup==0.3
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
requests==2.32.5
soupsieve==2.8.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.3.0
